+++ VerteilteSystemeProjekt/src/main/java/de/vsys/project/general/data/TargetTotalInfection.java
@@ -5,9 +5,9 @@ public class TargetTotalInfection {
     @author: David Rohrschneider
     This class is the template for the data which we return to the user
      */
-    private int targetTotalInfection;
+    private double targetTotalInfection;
 
-    public int getTargetTotalInfection() {
+    public double getTargetTotalInfection() {
         return targetTotalInfection;
     }
